menghapus file double dan memperbaiki template premium
