+++ frontend/src/App.jsx
@@ -1,3 +1,35 @@
+/* eslint-disable no-unused-vars */
+import React from 'react';
+import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
+import HomePageRoute from './Routes/HomePageRoute';
+import InvitationManager from './Routes/InvitationManager';
+
+const App = () => {
+  return (
+    <Router>
+      <Routes>
+        <Route path="/*" element={<HomePageRoute />} />
+        <Route path="/dashboard/*" element={<InvitationManager />} />
+      </Routes>
+    </Router>
+  );
+};
+
+export default App;
+
+
+
+
+
+
+
+
+
+
+
+
+
+
 // /* eslint-disable no-unused-vars */
 // import React from 'react';
 // import { BrowserRouter, Route, Routes } from 'react-router-dom';
@@ -41,23 +73,3 @@
 
 // export default App;
 
-
-
-/* eslint-disable no-unused-vars */
-import React from 'react';
-import { BrowserRouter as Router, Route, Routes } from "react-router-dom";
-import HomePageRoute from './Routes/HomePageRoute';
-import InvitationManager from './Routes/InvitationManager1';
-
-const App = () => {
-  return (
-    <Router>
-      <Routes>
-        <Route path="/*" element={<HomePageRoute />} />
-        <Route path="/dashboard/*" element={<InvitationManager />} />
-      </Routes>
-    </Router>
-  );
-};
-
-export default App;
